implement styledComponent to shorten code
